#123: little improvement for hibernate support

+++ mmm-util/mmm-util-core/src/main/java/net/sf/mmm/util/lang/base/datatype/adapter/hibernate/IntegerDatatypeUserType.java
@@ -24,7 +24,7 @@ public class IntegerDatatypeUserType<T extends SimpleDatatype<Integer>> extends 
    */
   public IntegerDatatypeUserType(Class<T> datatype) {
 
-    super(Types.NUMERIC, datatype, Integer.class);
+    super(Types.INTEGER, datatype, Integer.class);
   }
 
 }
